requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
boto3==1.36.21
botocore==1.36.21
CacheControl==0.14.2
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
cloudinary==1.36.0
colorama==0.4.6
cryptography==44.0.1
deprecation==2.1.0
distro==1.9.0
dj-database-url==2.3.0
Django==5.0.12
django-allauth==65.4.1
django-allauth-ui==1.5.3
django-ckeditor==6.7.3
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
django-filter==24.3
django-js-asset==3.1.2
django-storages==1.14.5
django-unfold==0.49.1
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
et_xmlfile==2.0.0
frozenlist==1.5.0
google-api-core==2.24.2
google-api-python-client==2.165.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.1
google-cloud-storage==3.1.0
google-crc32c==1.7.0
google-resumable-media==2.7.2
googleapis-common-protos==1.69.2
gotrue==2.11.4
grpcio==1.71.0
grpcio-status==1.71.0
gunicorn==23.0.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
jiter==0.9.0
jmespath==1.0.1
Markdown==3.7
msgpack==1.1.0
multidict==6.4.3
numpy==2.2.2
oauthlib==3.2.2
openai==1.68.2
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pluggy==1.5.0
postgrest==1.0.1
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
psycopg==3.2.4
psycopg-binary==3.2.4
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.9.0
PyMuPDF==1.25.5
pyparsing==3.2.3
pytest==8.3.5
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.1
PyYAML==6.0.2
razorpay==1.4.2
realtime==2.4.2
reportlab==4.4.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
s3transfer==0.11.2
setuptools==75.8.0
six==1.17.0
slippers==0.6.2
sniffio==1.3.1
sqlparse==0.5.3
storage3==0.11.3
StrEnum==0.4.15
stripe==11.5.0
supabase==2.15.0
supafunc==0.9.4
tqdm==4.67.1
typeguard==2.13.3
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
waitress==3.0.2
websockets==14.2
whitenoise==6.9.0
yarl==1.19.0
